moving server widgets to a widget dir
