+++ slacker/server/dashboard.js
@@ -1,4 +1,4 @@
-const Slacker = require('./slacker/slacker.js');
+const Slacker = require('./widget/slacker/slacker.js');
 
 const DASHBOARD_INTERVAL = 1000 * 30;
 const DASHBOARD_ERROR_INTERVAL = 1000 * 1;
@@ -7,10 +7,10 @@ function Dashboard(token) {
   this.slacker = new Slacker(token);
   this.index = 0;
   this.widgets = [
-    require('./ci-wall'),
-    require('./coverage'),
-    require('./stackoverflow'),
-    require('./reviews')
+    require('./widget/ci-wall'),
+    require('./widget/coverage'),
+    require('./widget/stackoverflow'),
+    require('./widget/reviews')
   ].concat(this.slacker.getRules());
 }
