+++ slacker/server/widget/ci-wall.js
