+++ slacker/server/widget/coverage.js
