+++ slacker/server/widget/slacker/gallery.js
