+++ slacker/server/widget/slacker/message-helper.js
