+++ slacker/server/widget/slacker/most-active-channel.js
