+++ slacker/server/widget/slacker/thanks.js
