+++ slacker/server/widget/stackoverflow.js
